==> notebooks/01_pull_price_data.ipynb <==
# Get prices

import pandas as pd
import numpy as np
import yfinance as yf

yf.__version__

## 1. Tesla

start = '2019-10-01'
end = '2024-09-27'

ticker = 'TSLA'

ticker = yf.Ticker(ticker)
prices = ticker.history(start=start, end=end)

len(prices)

prices['returns'] = prices['Close'].pct_change()
prices['log_returns'] = np.log(prices['Close']/prices['Close'].shift(1))
prices = prices.drop(columns=['Dividends', 'Stock Splits'])

prices

file_name = '../data/Tesla/tesla_prices.csv'
prices.to_csv(file_name)

## 1. S&P

start = '2019-09-30'
end = '2024-09-27'

ticker = 'SPY'

ticker = yf.Ticker(ticker)
prices = ticker.history(start=start, end=end)

len(prices)

prices['returns'] = prices['Close'].pct_change()
prices['log_returns'] = np.log(prices['Close']/prices['Close'].shift(1))
prices = prices.drop(columns=['Dividends', 'Stock Splits', 'Capital Gains'])

prices

file_name = '../data/S&P_market/s&p_prices.csv'
prices.to_csv(file_name)
