==> notebooks/02_pull_news_data.ipynb <==
## Retrieve news articles from NYT API
- To get API key: https://developer.nytimes.com/get-started

import os
import datetime
from dotenv import load_dotenv

import pandas as pd
import numpy as np

from pynytimes import NYTAPI

def get_news(nyt: NYTAPI,start_date: datetime.datetime, end_date: datetime.datetime, *, k=100, query=""):
    """
    get top k most relevent finance news headings on each day from NY times
    """
    res = []
    try:
        articles = nyt.article_search(
            query = query, 
            results = k,
            dates = {
                "begin": start_date,
                "end": end_date
            },
            options = {
                "sort": "relevance", 
                "sources": [
                    "New York Times",
                    "AP",
                    "Reuters",
                    "International Herald Tribune"
                ],
                # Only get information from these news desk
                "news_desk": [
                    "Business",
                    "Financial",
                    "Technology"
                ],
                # "body": [
                # ],
            }
        )
        
        for i in range(len(articles)):
            timestamp = articles[i]['pub_date']
            article_url = articles[i]['web_url']
            lead_paragraph = articles[i]['lead_paragraph']
            abstract = articles[i]['abstract']
            #print(articles[i])
            res.append({'timestamp': timestamp,
                        'article_url': article_url,
                        'lead_paragraph': lead_paragraph,
                        'abstract': abstract,
                        })
        
    except Exception as e:
        print(e)
        res.append({})
    
    df = pd.DataFrame(data=res)

    return df

#### Function to check if the news falls before or after market closes on each day. If it's after close, then the news should be categorised as a next day news.

def adjust_dates(df: pd.DataFrame) -> pd.DataFrame:
    threshold_time = pd.Timestamp('16:00:00').time()
    df['adjusted_date'] = df['timestamp'].apply(lambda x: x.date() if x.time() < threshold_time else (x + pd.Timedelta(days=1)).date())
    return df

#### Set NYTAPI

load_dotenv(dotenv_path='../.env')
NYT_API_KEY = os.getenv("NYT_API_KEY")

nyt = NYTAPI(NYT_API_KEY, parse_dates=True)

## 1. Get articles tesla 

start_date = datetime.datetime(2019, 9, 30)
end_date = datetime.datetime(2024, 9, 27)

k_results = int((end_date - start_date).days * 3)  # How many results to pull for. (x3 as a (conservative) estimate for the max articles that exist.)

article_df = get_news(nyt, start_date, end_date, k=k_results, query='Tesla') 
article_df = adjust_dates(article_df)
article_df = article_df.sort_values(by='timestamp')

article_df

filename = '../data/Tesla/tesla_headlines.csv'
article_df.to_csv(filename, index=False)

## 1. Get articles S&P (Market) 

start_date = datetime.datetime(2019, 9, 30)
end_date = datetime.datetime(2024, 9, 27)

k_results = int((end_date - start_date).days * 3)  # How many results to pull for. (x3 as a (conservative) estimate for the max articles that exist.)

article_df = get_news(nyt, start_date, end_date, k=k_results) 
article_df = adjust_dates(article_df)
article_df = article_df.sort_values(by='timestamp')

article_df

filename = '../data/S&P_market/s&p_headlines.csv'
article_df.to_csv(filename, index=False)
