==> notebooks/03a_openai_preprocessing.ipynb <==
import os

import pandas as pd
from dotenv import load_dotenv
from tqdm import tqdm
from openai import OpenAI

load_dotenv(dotenv_path='../.env')
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

client = OpenAI()

## Preprocessing with GPT

def parse_article_gpt(article, instructions):


    completion = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {"role": "system", "content": instructions},
            {
                "role": "user",
                "content": f"{article}"
            }
        ]
    )

    return completion.choices[0].message.content

### Test prompt

parse_article_gpt("Tesla reported record quarterly deliveries on Wednesday, putting the electric-car maker within reach of its goal for the year. But analysts remain skeptical that the company will show a profit for the period.","Evaluate, in one paragraph, how this piece of news would affect the Tesla stock price, while only using strictly either postiive or negative words before passing it into another sentiment scoring model.")

### Load Data

### Tesla

tesla_headlines = pd.read_csv('../data/Tesla/tesla_headlines.csv')

instructions_tesla = "Evaluate, in one paragraph, how this piece of news would affect the Tesla stock price, while only using strictly either postiive or negative words before passing it into another sentiment scoring model."


for i, row in tqdm(tesla_headlines.iterrows(), total=tesla_headlines.shape[0]):
    combined_string = f"{row['lead_paragraph']} \n {row['abstract']}"
    gpt_summary = parse_article_gpt(combined_string, instructions_tesla)
    tesla_headlines.loc[i, 'gpt_summary'] = gpt_summary


tesla_headlines.to_csv('../data/Tesla/tesla_gpt_summarised_sentiments.csv')

### S&P

spy_headlines = pd.read_csv('../data/S&P_market/s&p_headlines.csv')

instructions_spy = "Evaluate, in one paragraph, how this piece of news would affect the S&P index price, while only using strictly either postiive or negative words before passing it into another sentiment scoring model."


for i, row in tqdm(spy_headlines.iterrows(), total=spy_headlines.shape[0]):
    combined_string = f"{row['lead_paragraph']} \n {row['abstract']}"
    gpt_summary = parse_article_gpt(combined_string, instructions_spy)
    spy_headlines.loc[i, 'gpt_summary'] = gpt_summary


spy_headlines.to_csv('../data/S&P_market/s&p_gpt_summarised_sentiments.csv')
