==> notebooks/03b_sentiment_scoring_finbert.ipynb <==
## Sentiment scores generation using Finbert model
 - [Finbert github](https://github.com/ProsusAI/finBERT/tree/master)
 - [Huggingface link](https://huggingface.co/ProsusAI/finbert)

from transformers import AutoTokenizer, AutoModelForSequenceClassification
from torch import nn

import numpy as np
import pandas as pd

pd.set_option('display.max_colwidth', None)

tokenizer = AutoTokenizer.from_pretrained("ProsusAI/finbert")
model = AutoModelForSequenceClassification.from_pretrained("ProsusAI/finbert")

def generate_sentiment_values(text: str) -> list[float]:
    inputs = tokenizer(text, return_tensors='pt', padding=True)
    logits = model(**inputs)[0]

    # Convert logits to softmax probabilities.
    probabilities = nn.functional.softmax(logits, dim=-1)

    # The probability labels are 'positive', 'negative', 'neutral' (NEED TO CONFIRM)
    return probabilities.detach().numpy().reshape((3,))

### Sense check

test_string = 'Earnings reported are falling'
generate_sentiment_values(test_string)

## Finbert on Tesla news


1. Get sentiment for each article

tesla_df = pd.read_csv('../data/Tesla/tesla_gpt_summarised_sentiments.csv')
#df = df[['timestamp','article_url','lead_paragraph','abstract','adjusted_date']]

tesla_df

for i, row in tesla_df.iterrows():
    output = generate_sentiment_values(row['gpt_summary'])
    preamble_tesla = "Evaluate the above news on Tesla's stock price."
    combined_str = f"""<news>{row['gpt_summary']}</news> <instructions>{preamble_tesla}</instructions> """    
    output = generate_sentiment_values(combined_str)
    tesla_df.at[i, 'pos_sentiment'] = output[0]
    tesla_df.at[i, 'neg_sentiment'] = output[1]
    tesla_df.at[i, 'neutral_sentiment'] = output[2]

Save to the same file first

tesla_df

tesla_df.to_csv('../data/Tesla/tesla_gpt_summarised_sentiments.csv', index=False)

## Finbert on S&P news


1. Get sentiment for each article

spy_df = pd.read_csv('../data/S&P_market/s&p_gpt_summarised_sentiments.csv')
#df = df[['timestamp','article_url','lead_paragraph','abstract','adjusted_date']]

spy_df

for i, row in spy_df.iterrows():
    output = generate_sentiment_values(row['gpt_summary'])
    preamble_spy = "Evaluate the above news on S&P stock index."
    combined_str = f"""<news>{row['gpt_summary']}</news> <instructions>{preamble_spy}</instructions> """    
    output = generate_sentiment_values(combined_str)
    spy_df.at[i, 'pos_sentiment'] = output[0]
    spy_df.at[i, 'neg_sentiment'] = output[1]
    spy_df.at[i, 'neutral_sentiment'] = output[2]

spy_df

tesla_df.to_csv('../data/S&P_market/s&p_gpt_summarised_sentiments.csv', index=False)
